컨버터 수정, 배틀테그상세 페이지 수정, 하단에 히어로 목록 추가 목록의 디자인은 공홈에 나오는 데로 만들 예정(ppt5page 참고)

+++ Diablo3Hub/Diablo3Hub/Models/CareerProfile.cs
@@ -26,10 +26,10 @@ namespace Diablo3Hub.Models
         public Hero[] Heroes { get; set; }
 
         [JsonProperty(PropertyName = "lastHeroPlayed")]
-        public int LastHeroPlayed { get; set; }
+        public long LastHeroPlayed { get; set; }
 
         [JsonProperty(PropertyName = "lastUpdated")]
-        public int LastUpdated { get; set; }
+        public long LastUpdated { get; set; }
 
         [JsonProperty(PropertyName = "kills")]
         public Kills Kills { get; set; }
@@ -246,6 +246,6 @@ namespace Diablo3Hub.Models
         public bool Dead { get; set; }
 
         [JsonProperty(PropertyName = "lastupdated")]
-        public int Lastupdated { get; set; }
+        public long Lastupdated { get; set; }
     }
 }
